# .futurxlab/api-specification.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    InKnowing is an AI-powered book dialogue platform that enables users to have conversations with books and book characters.

    ## Business Logic Conservation
    Based on the principle of business logic conservation, this API documentation maintains complete mapping between:
    - User Journey → API Endpoints
    - State Transitions → API Operations
    - Sequence Interactions → API Calls
    - Business Rules → API Responses

    ## Key Features
    - Question-driven and book-driven discovery
    - AI-powered dialogues with books and characters
    - User book upload with vectorization
    - Multi-tier membership system
    - Admin management dashboard

    ## Authentication
    The API uses JWT Bearer tokens for authentication. Tokens are obtained through the login endpoints and must be included in the Authorization header for protected endpoints.

    ## Rate Limiting
    - Free users: 20 requests per day
    - Basic members: 200 requests per month
    - Premium members: 500 requests per month
    - Super members: 1000 requests per month
  version: 1.0.0
  contact:
    name: InKnowing Support
    email: support@inknowing.ai
    url: https://inknowing.ai/support
  license:
    name: Proprietary
    url: https://inknowing.ai/terms
servers:
  - url: https://api.inknowing.ai/v1
    description: Production server
  - url: https://staging-api.inknowing.ai/v1
    description: Staging server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: |
        Register a new user account with phone number or WeChat
        Business Logic: User Journey Start → Registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PhoneRegistration'
                - $ref: '#/components/schemas/WeChatRegistration'
            examples:
              phone:
                value:
                  type: phone
                  phone: "13800138000"
                  code: "123456"
              wechat:
                value:
                  type: wechat
                  code: "wx_auth_code_123"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: |
        Login with phone/password or WeChat authorization
        Business Logic: User Authentication State Transition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/PhoneLogin'
                - $ref: '#/components/schemas/WeChatLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange refresh token for new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/verify-code:
    post:
      tags:
        - Authentication
      summary: Send verification code
      description: Send SMS verification code for phone registration/login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phone
              properties:
                phone:
                  type: string
                  pattern: '^1[3-9]\d{9}$'
      responses:
        '200':
          description: Code sent successfully
        '429':
          description: Too many requests

  # ==================== Users ====================
...
  /books:
    get:
      tags:
        - Books
      summary: List books
      description: Get paginated list of available books
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [business, psychology, fiction, science, history, philosophy]
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, newest, most_discussed]
            default: popular
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Book list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookList'
  /books/popular:
    get:
      tags:
        - Books
      summary: Get popular books
      description: Get list of most popular books based on dialogue count
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
            default: week
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 10
      responses:
        '200':
          description: Popular books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookList'
  /books/{bookId}:
    get:
      tags:
        - Books
      summary: Get book details
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /books/{bookId}/characters:
    get:
      tags:
        - Books
      summary: Get book characters
      description: Get list of available dialogue characters for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Character list
          content:
            application/json:
              schema:
                type: object
                properties:
                  characters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Character'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== Dialogues ====================
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        phone:
          type: string
          nullable: true
        wechat_openid:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
        nickname:
          type: string
        membership:
          $ref: '#/components/schemas/MembershipType'
        points:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: Bearer
        expires_in:
          type: integer
          description: Token expiry time in seconds
        user:
          $ref: '#/components/schemas/User'
...
    PhoneRegistration:
      type: object
      required:
        - type
        - phone
        - code
      properties:
        type:
          type: string
          enum: [phone]
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        code:
          type: string
          pattern: '^\d{6}$'
        password:
          type: string
          minLength: 6
    WeChatRegistration:
      type: object
      required:
        - type
        - code
      properties:
        type:
          type: string
          enum: [wechat]
        code:
          type: string
          description: WeChat authorization code
    PhoneLogin:
      type: object
      required:
        - type
        - phone
      properties:
        type:
          type: string
          enum: [phone]
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$'
        password:
          type: string
          minLength: 6
        code:
          type: string
          pattern: '^\d{6}$'
          description: SMS code for code-based login
    WeChatLogin:
      type: object
      required:
        - type
        - code
      properties:
        type:
          type: string
          enum: [wechat]
        code:
          type: string
          description: WeChat authorization code

    # ==================== Membership Schemas ====================
    MembershipType:
      type: string
      enum: [free, basic, premium, super]
...
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        cover:
          type: string
          nullable: true
        category:
          type: string
        description:
          type: string
        dialogue_count:
          type: integer
        rating:
          type: number
          minimum: 0
          maximum: 5
        created_at:
          type: string
          format: date-time
    BookDetail:
      allOf:
        - $ref: '#/components/schemas/Book'
        - type: object
          properties:
            type:
              type: string
              enum: [ai_known, vectorized]
            chapters:
              type: integer
            estimated_reading_time:
              type: integer
              description: Estimated reading time in minutes
            characters:
              type: array
              items:
                $ref: '#/components/schemas/Character'
            tags:
              type: array
              items:
                type: string
            uploader:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                nickname:
                  type: string
...
    BookList:
      type: object
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Character:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        alias:
          type: array
          items:
            type: string
        description:
          type: string
        personality:
          type: string
        dialogue_count:
          type: integer
        enabled:
          type: boolean
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication

    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

# API Response Standards
...